requirements.txt
fastapi~=0.122.0
celery~=5.5.3
requests~=2.32.5
moviepy~=2.2.1
pydantic~=2.12.4
uvicorn
redis
pydantic
google-cloud-storage
python-multipart
python-decouple